+++ FormalLanguage/CFG/Outside.hs
@@ -31,7 +31,7 @@ outsideFromInside g
   -- (linear grammars are context-free as well).
   -- not $ isContextFree g = Nothing
   | otherwise = Just $ Grammar {..}
-  where _outside     = Outside (g^.grammarName)
+  where _outside     = Outside g
         _rules       = S.fromList $ epsrule : (concatMap genOutsideRules $ g^..rules.folded)
         _grammarName = "" -- will be set in the parser
         _params      = g^.params
@@ -85,7 +85,7 @@ outsideSymb = over (getSymbolList . traverse . name . getSteName) (++"'")
 -- | 
 
 genEpsilonRule :: [AttributeFunction] -> Symbol -> Rule
-genEpsilonRule epsfun s = Rule (outsideSymb s) epsfun [(Symbol $ replicate (length $ s^.getSymbolList) Epsilon)]
+genEpsilonRule epsfun s = Rule (outsideSymb s) epsfun [(Symbol $ replicate (length $ s^.getSymbolList) $ Epsilon Global)]
 
 -- |