+++ FormalLanguage/CFG/TH/Internal.hs
@@ -7,7 +7,9 @@ import qualified Language.Haskell.TH.Lib as TH
 import           Language.Haskell.TH
 
 
-
+#if MIN_VERSION_base(4,10,0)
+dataD ctxt tc tvs cons = TH.dataD ctxt tc tvs Nothing cons []
+#else
 #if MIN_VERSION_base(4,9,0)
 --dataD :: CxtQ -> Name -> [TyVarBndr] -> [ConQ] -> CxtQ -> DecQ
 --
@@ -16,5 +18,6 @@ import           Language.Haskell.TH
 dataD ctxt tc tvs cons = TH.dataD ctxt tc tvs Nothing cons (return [])
 #else
 dataD ctxt tc tvs cons = TH.dataD ctxt tc tvs cons []
+#endif
 #endif