+++ src/NeedlemanWunsch.hs
@@ -9,14 +9,15 @@ import           Control.Monad.ST
 import           Data.Char (toUpper,toLower)
 import           Data.List (take)
 import           Data.Vector.Fusion.Util
+import           Data.Vector.Unboxed (Vector)
 import           Language.Haskell.TH
 import           Language.Haskell.TH.Syntax
 import qualified Data.Vector.Fusion.Stream.Monadic as SM
 import qualified Data.Vector.Unboxed as VU
-import           Data.Vector.Unboxed (Vector)
+import           System.Environment (getArgs)
 import           Text.Printf
 
-import           ADP.Fusion
+import           ADP.Fusion.PointL
 import           Data.PrimitiveArray as PA hiding (map,toList)
 
 import           FormalLanguage.CFG
@@ -38,7 +39,12 @@ S: [X,X]
 [X,X] -> delin <<< [X,X] [c,-]
 //
 
+Outside: Labolg
+Source: Global
+//
+
 Emit: Global
+Emit: Labolg
 |]
 
 
@@ -58,47 +64,90 @@ score = SigGlobal
 {-# INLINE score #-}
 
 
--- | 
+-- | Pretty-print the optimal alignment.
 --
--- NOTE The alignment needs to be reversed to print out.
+-- NOTE The alignment needs to be reversed to print out. @x ++ [a]@ is
+-- @O(n^2)@, while @reverse@ on the final string with @a:x@ is @O(n)@. For
+-- long inputs (say 1000nt) prettyprinting will otherwise take *longer*
+-- than scoring.
 
 pretty :: Monad m => SigGlobal m (String,String) [(String,String)] Char Char
 pretty = SigGlobal
-  { done  = \       (Z:.():.()) -> ("","")
-  , align = \ (x,y) (Z:.a :.b ) -> (x ++ [a] ,y ++ [b]) 
-  , indel = \ (x,y) (Z:.():.b ) -> (x ++ "-" ,y ++ [b]) 
-  , delin = \ (x,y) (Z:.a :.()) -> (x ++ [a] ,y ++ "-") 
+  { done  = \       (Z:.():.()) -> (""   ,""   )
+  , align = \ (x,y) (Z:.a :.b ) -> (a:x  , b:y )
+  , indel = \ (x,y) (Z:.():.b ) -> ('-':x, b:y )
+  , delin = \ (x,y) (Z:.a :.()) -> (a:x  ,'-':y)
   , h     = SM.toList
   }
 {-# Inline pretty #-}
 
-runNeedlemanWunsch :: Int -> String -> String -> (Int,[(String,String)])
-runNeedlemanWunsch k i1' i2' = (d, take k . unId $ axiom b) where
+runNeedlemanWunschF :: Int -> String -> String -> (Int,[(String,String)],PerfCounter)
+runNeedlemanWunschF k i1' i2' = (d, take k . unId $ axiom b, perf) where
   i1 = VU.fromList i1'
   i2 = VU.fromList i2'
-  !(Z:.t) = runNeedlemanWunschForward i1 i2
+  Mutated (Z:.t) perf eachPerf = runNeedlemanWunschForward i1 i2
   d = unId $ axiom t
   !(Z:.b) = gGlobal (score <|| pretty) (toBacktrack t (undefined :: Id a -> Id a)) (chr i1) (chr i2)
-{-# NoInline runNeedlemanWunsch #-}
+{-# NoInline runNeedlemanWunschF #-}
+
+runNeedlemanWunschB :: Int -> String -> String -> (Int,[(String,String)],PerfCounter)
+runNeedlemanWunschB k i1' i2' = (d, take k . unId $ axiom b, perf) where
+  i1 = VU.fromList i1'
+  i2 = VU.fromList i2'
+  Mutated (Z:.t) perf eachPerf = runNeedlemanWunschBackward i1 i2
+  d = unId $ axiom t
+  !(Z:.b) = gGlobal (score <|| pretty) (toBacktrack t (undefined :: Id a -> Id a)) (chr i1) (chr i2)
+{-# NoInline runNeedlemanWunschB #-}
 
 -- | Decoupling the forward phase for CORE observation.
 
-runNeedlemanWunschForward :: Vector Char -> Vector Char -> Z:.(TwITbl Id Unboxed (Z:.EmptyOk:.EmptyOk) (Z:.PointL I:.PointL I) Int)
-runNeedlemanWunschForward i1 i2 = let n1 = VU.length i1; n2 = VU.length i2 in mutateTablesDefault $
-  gGlobal score
-    (ITbl 0 0 (Z:.EmptyOk:.EmptyOk) (PA.fromAssocs (Z:.PointL 0:.PointL 0) (Z:.PointL n1:.PointL n2) (-999999) []))
+runNeedlemanWunschForward
+  ∷ Vector Char
+  → Vector Char
+  → Mutated (Z:.TwITbl _ _ Id Unboxed (Z:.EmptyOk:.EmptyOk) (Z:.PointL I:.PointL I) Int)
+runNeedlemanWunschForward i1 i2 = runST $ do
+  arr ← newWithPA (ZZ:..LtPointL n1:..LtPointL n2) (-999999)
+  ts ← fillTables $ gGlobal score
+    (ITbl @0 @0 (Z:.EmptyOk:.EmptyOk) arr)
     (chr i1) (chr i2)
+  return ts
+  where n1 = VU.length i1
+        n2 = VU.length i2
 {-# NoInline runNeedlemanWunschForward #-}
 
+-- | Decoupling the forward phase for CORE observation.
+
+runNeedlemanWunschBackward
+  ∷ Vector Char
+  → Vector Char
+  → Mutated (Z:.TwITbl _ _ Id Unboxed (Z:.EmptyOk:.EmptyOk) (Z:.PointL O:.PointL O) Int)
+runNeedlemanWunschBackward i1 i2 = runST $ do
+  arr ← newWithPA (ZZ:..LtPointL n1:..LtPointL n2) (-999999)
+  ts ← fillTables $ gLabolg score
+    (ITbl @0 @0 (Z:.EmptyOk:.EmptyOk) arr)
+    (chr i1) (chr i2)
+  return ts
+  where n1 = VU.length i1
+        n2 = VU.length i2
+{-# NoInline runNeedlemanWunschBackward #-}
+
+
+
 main = do
+  as <- getArgs
+  let k = if null as then 1 else read $ head as
   ls <- lines <$> getContents
   let eats [] = return ()
       eats [x] = return ()
       eats (a:b:xs) = do
         putStrLn a
         putStrLn b
-        let (k,ys) = runNeedlemanWunsch 1 a b
-        forM_ ys $ \(y1,y2) -> printf "%s %5d\n%s\n" y1 k y2
+        let (sF,ysF,perfF) = runNeedlemanWunschF k a b
+        let (sB,ysB,perfB) = runNeedlemanWunschB k a b
+        forM_ ysF $ \(y1,y2) -> printf "%s\n%s\n%5d (backward: %5d)\n" (reverse y1) (reverse y2) sF sB
+        when (k==0) $ print sF
+        print perfF
+        print perfB
         eats xs
   eats ls