+++ tests/properties.hs
@@ -1,14 +1,8 @@
 
 module Main where
 
-import Test.Tasty
-import Test.Tasty.TH
-
 
 
 main :: IO ()
-main = do
-  defaultMain $ testGroup ""
-    [
-    ]
+main = return ()