+++ changelog.md
@@ -1,7 +1,9 @@
-0.3.1.2
+0.3.2.0
 -------
 
-- removed superfluous instance which prevents compilation
+- parsers 0.12.5 added mtl instances for Unlined, etc. We remove our own
+  instances and require 0.12.5 as the new minimal version
+- bang -> notStrict (GHC 8 only)
 
 0.3.1.1
 -------